requirements.txt
absl-py==2.0.0
apache-beam==2.51.0
attrs==23.1.0
bidict==0.22.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
eventlet==0.33.3
fastavro==1.8.4
fasteners==0.19
Flask==3.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.43.1
google-api-core==2.12.0
google-apitools==0.5.31
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.12.0
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.21.0
google-cloud-core==2.3.3
google-cloud-datastore==2.18.0
google-cloud-dlp==3.12.3
google-cloud-language==2.11.1
google-cloud-pubsub==2.18.4
google-cloud-pubsublite==1.8.3
google-cloud-recommendations-ai==0.10.5
google-cloud-resource-manager==1.10.4
google-cloud-spanner==3.40.1
google-cloud-storage==2.12.0
google-cloud-videointelligence==2.11.4
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mediapipe==0.10.7
numpy==1.24.4
oauth2client==4.1.3
objsize==0.6.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
orjson==3.9.9
overrides==6.5.0
packaging==23.2
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
rsa==4.9
shapely==2.0.2
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.7
Werkzeug==3.0.0
wsproto==1.2.0
zstandard==0.21.0
